+++ wms/src/main/java/com/shinlogis/wms/inoutbound/inbound/repository/ReceiptDAO.java
@@ -1,4 +1,4 @@
-package com.shinlogis.wms.inbound.repository;
+package com.shinlogis.wms.inoutbound.inbound.repository;
 
 import java.sql.Connection;
 import java.sql.Date;
