+++ wms/src/main/java/com/shinlogis/wms/inoutbound/outbound/view/OutboundDetailModel.java
@@ -1,0 +1,5 @@
+package com.shinlogis.wms.inoutbound.outbound.view;
+
+public class OutboundDetailModel {
+
+}
