--- wms/src/main/java/com/shinlogis/wms/outbound/view/OutboundDetailModel.java
@@ -1,5 +1,0 @@
-package com.shinlogis.wms.outbound.view;
-
-public class OutboundDetailModel {
-
-}
